+++ assets/js/paper-dashboard.js
@@ -15,18 +15,17 @@
 
  */
 
-(function() {
-  isWindows = navigator.platform.indexOf('Win') > -1 ? true : false;
+(function () {
+  isWindows = navigator.platform.indexOf("Win") > -1 ? true : false;
 
   if (isWindows) {
     // if we are on windows OS we activate the perfectScrollbar function
-    var ps = new PerfectScrollbar('.sidebar');
-    var ps1 = new PerfectScrollbar('.sidebar-wrapper');
-    var ps2 = new PerfectScrollbar('.main-panel');
-    $('html').addClass('perfect-scrollbar-on');
-
+    var ps = new PerfectScrollbar(".sidebar");
+    var ps1 = new PerfectScrollbar(".sidebar-wrapper");
+    var ps2 = new PerfectScrollbar(".main-panel");
+    $("html").addClass("perfect-scrollbar-on");
   } else {
-    $('html').addClass('perfect-scrollbar-off');
+    $("html").addClass("perfect-scrollbar-off");
   }
 })();
 
@@ -39,150 +38,169 @@ backgroundOrange = false;
 sidebar_mini_active = false;
 toggle_initialized = false;
 
-seq = 0, delays = 80, durations = 500;
-seq2 = 0, delays2 = 80, durations2 = 500;
+(seq = 0), (delays = 80), (durations = 500);
+(seq2 = 0), (delays2 = 80), (durations2 = 500);
 
-$(document).ready(function() {
-
-  if ($('.full-screen-map').length == 0 && $('.bd-docs').length == 0) {
+$(document).ready(function () {
+  if ($(".full-screen-map").length == 0 && $(".bd-docs").length == 0) {
     // On click navbar-collapse the menu will be white not transparent
-    $('.collapse').on('show.bs.collapse', function() {
-      $(this).closest('.navbar').removeClass('navbar-transparent').addClass('bg-white');
-    }).on('hide.bs.collapse', function() {
-      $(this).closest('.navbar').addClass('navbar-transparent').removeClass('bg-white');
-    });
+    $(".collapse")
+      .on("show.bs.collapse", function () {
+        $(this)
+          .closest(".navbar")
+          .removeClass("navbar-transparent")
+          .addClass("bg-white");
+      })
+      .on("hide.bs.collapse", function () {
+        $(this)
+          .closest(".navbar")
+          .addClass("navbar-transparent")
+          .removeClass("bg-white");
+      });
   }
 
   paperDashboard.initMinimizeSidebar();
 
-  $navbar = $('.navbar[color-on-scroll]');
-  scroll_distance = $navbar.attr('color-on-scroll') || 500;
+  $navbar = $(".navbar[color-on-scroll]");
+  scroll_distance = $navbar.attr("color-on-scroll") || 500;
 
   // Check if we have the class "navbar-color-on-scroll" then add the function to remove the class "navbar-transparent" so it will transform to a plain color.
-  if ($('.navbar[color-on-scroll]').length != 0) {
+  if ($(".navbar[color-on-scroll]").length != 0) {
     paperDashboard.checkScrollForTransparentNavbar();
-    $(window).on('scroll', paperDashboard.checkScrollForTransparentNavbar)
+    $(window).on("scroll", paperDashboard.checkScrollForTransparentNavbar);
   }
 
-  $('.form-control').on("focus", function() {
-    $(this).parent('.input-group').addClass("input-group-focus");
-  }).on("blur", function() {
-    $(this).parent(".input-group").removeClass("input-group-focus");
-  });
+  $(".form-control")
+    .on("focus", function () {
+      $(this).parent(".input-group").addClass("input-group-focus");
+    })
+    .on("blur", function () {
+      $(this).parent(".input-group").removeClass("input-group-focus");
+    });
 
   // Activate bootstrapSwitch
-  $('.bootstrap-switch').each(function() {
+  $(".bootstrap-switch").each(function () {
     $this = $(this);
-    data_on_label = $this.data('on-label') || '';
-    data_off_label = $this.data('off-label') || '';
+    data_on_label = $this.data("on-label") || "";
+    data_off_label = $this.data("off-label") || "";
 
     $this.bootstrapSwitch({
       onText: data_on_label,
-      offText: data_off_label
+      offText: data_off_label,
     });
   });
 });
 
-$(document).on('click', '.navbar-toggle', function() {
+$(document).on("click", ".navbar-toggle", function () {
   $toggle = $(this);
 
   if (paperDashboard.misc.navbar_menu_visible == 1) {
-    $('html').removeClass('nav-open');
+    $("html").removeClass("nav-open");
     paperDashboard.misc.navbar_menu_visible = 0;
-    setTimeout(function() {
-      $toggle.removeClass('toggled');
-      $('#bodyClick').remove();
+    setTimeout(function () {
+      $toggle.removeClass("toggled");
+      $("#bodyClick").remove();
     }, 550);
-
   } else {
-    setTimeout(function() {
-      $toggle.addClass('toggled');
+    setTimeout(function () {
+      $toggle.addClass("toggled");
     }, 580);
 
     div = '<div id="bodyClick"></div>';
-    $(div).appendTo('body').click(function() {
-      $('html').removeClass('nav-open');
-      paperDashboard.misc.navbar_menu_visible = 0;
-      setTimeout(function() {
-        $toggle.removeClass('toggled');
-        $('#bodyClick').remove();
-      }, 550);
-    });
+    $(div)
+      .appendTo("body")
+      .click(function () {
+        $("html").removeClass("nav-open");
+        paperDashboard.misc.navbar_menu_visible = 0;
+        setTimeout(function () {
+          $toggle.removeClass("toggled");
+          $("#bodyClick").remove();
+        }, 550);
+      });
 
-    $('html').addClass('nav-open');
+    $("html").addClass("nav-open");
     paperDashboard.misc.navbar_menu_visible = 1;
   }
 });
 
-$(window).resize(function() {
+$(window).resize(function () {
   // reset the seq for charts drawing animations
   seq = seq2 = 0;
 
-  if ($('.full-screen-map').length == 0 && $('.bd-docs').length == 0) {
-    $navbar = $('.navbar');
-    isExpanded = $('.navbar').find('[data-toggle="collapse"]').attr("aria-expanded");
-    if ($navbar.hasClass('bg-white') && $(window).width() > 991) {
-      $navbar.removeClass('bg-white').addClass('navbar-transparent');
-    } else if ($navbar.hasClass('navbar-transparent') && $(window).width() < 991 && isExpanded != "false") {
-      $navbar.addClass('bg-white').removeClass('navbar-transparent');
+  if ($(".full-screen-map").length == 0 && $(".bd-docs").length == 0) {
+    $navbar = $(".navbar");
+    isExpanded = $(".navbar")
+      .find('[data-toggle="collapse"]')
+      .attr("aria-expanded");
+    if ($navbar.hasClass("bg-white") && $(window).width() > 991) {
+      $navbar.removeClass("bg-white").addClass("navbar-transparent");
+    } else if (
+      $navbar.hasClass("navbar-transparent") &&
+      $(window).width() < 991 &&
+      isExpanded != "false"
+    ) {
+      $navbar.addClass("bg-white").removeClass("navbar-transparent");
     }
   }
 });
 
 paperDashboard = {
   misc: {
-    navbar_menu_visible: 0
+    navbar_menu_visible: 0,
   },
 
-  initMinimizeSidebar: function() {
-    if ($('.sidebar-mini').length != 0) {
+  initMinimizeSidebar: function () {
+    if ($(".sidebar-mini").length != 0) {
       sidebar_mini_active = true;
     }
 
-    $('#minimizeSidebar').click(function() {
+    $("#minimizeSidebar").click(function () {
       var $btn = $(this);
 
       if (sidebar_mini_active == true) {
-        $('body').addClass('sidebar-mini');
+        $("body").addClass("sidebar-mini");
         sidebar_mini_active = true;
-        paperDashboard.showSidebarMessage('Sidebar mini activated...');
+        paperDashboard.showSidebarMessage("Sidebar mini activated...");
       } else {
-        $('body').removeClass('sidebar-mini');
+        $("body").removeClass("sidebar-mini");
         sidebar_mini_active = false;
-        paperDashboard.showSidebarMessage('Sidebar mini deactivated...');
+        paperDashboard.showSidebarMessage("Sidebar mini deactivated...");
       }
 
       // we simulate the window Resize so the charts will get updated in realtime.
-      var simulateWindowResize = setInterval(function() {
-        window.dispatchEvent(new Event('resize'));
+      var simulateWindowResize = setInterval(function () {
+        window.dispatchEvent(new Event("resize"));
       }, 180);
 
       // we stop the simulation of Window Resize after the animations are completed
-      setTimeout(function() {
+      setTimeout(function () {
         clearInterval(simulateWindowResize);
       }, 1000);
     });
   },
 
-  showSidebarMessage: function(message) {
+  showSidebarMessage: function (message) {
     try {
-      $.notify({
-        icon: "now-ui-icons ui-1_bell-53",
-        message: message
-      }, {
-        type: 'info',
-        timer: 4000,
-        placement: {
-          from: 'top',
-          align: 'right'
+      $.notify(
+        {
+          icon: "now-ui-icons ui-1_bell-53",
+          message: message,
+        },
+        {
+          type: "info",
+          timer: 4000,
+          placement: {
+            from: "top",
+            align: "right",
+          },
         }
-      });
+      );
     } catch (e) {
-      console.log('Notify library is missing, please make sure you have the notifications library added.');
+      console.log(
+        "Notify library is missing, please make sure you have the notifications library added."
+      );
     }
-
-  }
-
+  },
 };
 
 function hexToRGB(hex, alpha) {