+++ python/noverview.py
@@ -4,13 +4,17 @@ import io
 import numpy as np
 from datetime import date, timedelta
 
-dates = date.today()-timedelta(days=1)
-url = "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/"+dates.strftime('%m-%d-%Y')+".csv"
+dates = date.today() - timedelta(days=1)
+url = (
+    "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/"
+    + dates.strftime("%m-%d-%Y")
+    + ".csv"
+)
 download = requests.get(url).content
 
-df = pd.read_csv(io.StringIO(download.decode('utf-8')))
+df = pd.read_csv(io.StringIO(download.decode("utf-8")))
 
-htmlBase = '''
+htmlBase = """
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -30,25 +34,44 @@ htmlBase = '''
 </style>
 </head>
 <body>
-'''
+"""
 
-with open('../us/noverview.html', 'w') as overview:
+with open("../us/noverview.html", "w") as overview:
     overview.write(htmlBase)
 
-with open('../us/noverview.html', 'a') as overview:
-    deaths = '{:,}'.format(np.sum(np.array(df['Deaths'].to_list())))
-    cases = '{:,}'.format(np.sum(np.array(df['Confirmed'].to_list())))
+with open("../us/noverview.html", "a") as overview:
+    deaths = "{:,}".format(np.sum(np.array(df["Deaths"].to_list())))
+    cases = "{:,}".format(np.sum(np.array(df["Confirmed"].to_list())))
     try:
-        recovered = '{:,}'.format(np.sum(np.array(df['Confirmed'].to_list()))-np.sum(np.array(list(map(int, df['Active'].to_list())))))
+        recovered = "{:,}".format(
+            np.sum(np.array(df["Confirmed"].to_list()))
+            - np.sum(np.array(list(map(int, df["Active"].to_list()))))
+        )
     except:
-        recovered = 'No current data'
+        recovered = "No current data"
     try:
-        active = '{:,}'.format(np.sum(np.array(list(map(int, df['Active'].to_list())))))
+        active = "{:,}".format(np.sum(np.array(list(map(int, df["Active"].to_list())))))
     except:
-        active = 'No current data'
+        active = "No current data"
 
-    overview.write('<h6 class=\'text-primary\' align=\'center\'>Total Cases:   '+str(cases)+'</h6><br>')
-    overview.write('<h6 class=\'text-primary\' align=\'center\'>Total Deaths: '+str(deaths)+'</h6><br>')
-    overview.write('<h6 class=\'text-primary\' align=\'center\'>Total Recovered: '+str(recovered)+'</h6><br>')
-    overview.write('<h6 class=\'text-primary\' align=\'center\'>Total Active: '+str(active)+'</h6><br>')
-    overview.write('</body></html>')
+    overview.write(
+        "<h6 class='text-primary' align='center'>Total Cases:   "
+        + str(cases)
+        + "</h6><br>"
+    )
+    overview.write(
+        "<h6 class='text-primary' align='center'>Total Deaths: "
+        + str(deaths)
+        + "</h6><br>"
+    )
+    overview.write(
+        "<h6 class='text-primary' align='center'>Total Recovered: "
+        + str(recovered)
+        + "</h6><br>"
+    )
+    overview.write(
+        "<h6 class='text-primary' align='center'>Total Active: "
+        + str(active)
+        + "</h6><br>"
+    )
+    overview.write("</body></html>")