+++ python/tables.py
@@ -3,22 +3,26 @@ import requests
 import io
 from datetime import date, timedelta
 
-dates = date.today()-timedelta(days=1)
-url = "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/"+dates.strftime('%m-%d-%Y')+".csv"
+dates = date.today() - timedelta(days=1)
+url = (
+    "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/"
+    + dates.strftime("%m-%d-%Y")
+    + ".csv"
+)
 download = requests.get(url).content
 url2 = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/us_state_vaccinations.csv"
 download2 = requests.get(url2).content
 
-df = pd.read_csv(io.StringIO(download.decode('utf-8')))
-df2 = pd.read_csv(io.StringIO(download2.decode('utf-8')))
+df = pd.read_csv(io.StringIO(download.decode("utf-8")))
+df2 = pd.read_csv(io.StringIO(download2.decode("utf-8")))
 
 vaccinations = []
 
 for x in range(0, len(df2["date"])):
-    if df2["date"][x] == dates.strftime('%Y-%m-%d'):
+    if df2["date"][x] == dates.strftime("%Y-%m-%d"):
         vaccinations.append([df2["location"][x], df2["total_vaccinations"][x]])
 
-htmlBase = '''
+htmlBase = """
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -38,13 +42,14 @@ htmlBase = '''
 </style>
 </head>
 <body>
-'''
+"""
 
-with open('../us/table.html', 'w') as table:
+with open("../us/table.html", "w") as table:
     table.write(htmlBase)
 
-with open('../us/table.html', 'a') as table:
-    table.write('''
+with open("../us/table.html", "a") as table:
+    table.write(
+        """
     <table align="center" class='table' width='100%'>
         <tr class='text-primary'>
             <th>Last Updated</th> 
@@ -53,30 +58,45 @@ with open('../us/table.html', 'a') as table:
             <th>Deaths</th>
             <th>Vaccinations</th>
         </tr>
-    ''')
+    """
+    )
     for x in range(0, 58):
-        '''
+        """
         active = 'N/A'
         recovered = 'N/A'
-        
+
         if df["Active"][x] == df["Active"][x]:
             active = int(df["Active"][x])
             recovered = int(df["Confirmed"][x]) - active
-        ''' 
+        """
         for z in vaccinations:
             if z[0] == df["Province_State"][x]:
                 vax_data = z[1]
 
-        table.write('''
+        table.write(
+            """
         <tr>
-            <th>'''+str(df["Last_Update"][x])+'''</th> 
-            <th>'''+str(df["Province_State"][x])+'''</th> 
-            <th>'''+'{:,}'.format(df["Confirmed"][x])+'''</th> 
-            <th>'''+'{:,}'.format(df["Deaths"][x])+'''</th> 
-            <th>'''+'{:,}'.format(int(vax_data))+'''</th>
+            <th>"""
+            + str(df["Last_Update"][x])
+            + """</th> 
+            <th>"""
+            + str(df["Province_State"][x])
+            + """</th> 
+            <th>"""
+            + "{:,}".format(df["Confirmed"][x])
+            + """</th> 
+            <th>"""
+            + "{:,}".format(df["Deaths"][x])
+            + """</th> 
+            <th>"""
+            + "{:,}".format(int(vax_data))
+            + """</th>
         <tr>
-        ''')
+        """
+        )
 
-    table.write('''</table>
+    table.write(
+        """</table>
     </body>
-</html>''')
+</html>"""
+    )