+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
+
 View all updates to the Dashboard here. Releases are structured `vx.x.x-[Commits since base version released]`. For example, after 30 commits of the base `v1.0.0` release, the version will be `v1.0.0-30`.
 
 ## v1.0.0-alpha
+
 Initial Release with:
- - US Stats and graphs
- - Uses Paper Dashboard 2 HTML template by Creative Tim
+
+- US Stats and graphs
+- Uses Paper Dashboard 2 HTML template by Creative Tim